+++ .eslintrc.cjs
@@ -1,5 +1,5 @@
 module.exports = {
   root: true,
   extends: ['@ionic/eslint-config/recommended'],
-  ignorePatterns: ['dist/', 'node_modules/', 'example/'],
+  ignorePatterns: ['dist/', 'node_modules/', 'example/', 'demo/', 'test/'],
 };